Restyle and reorganize components
